% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HeatmapDE.R
\name{heatmap.sig.genes}
\alias{heatmap.sig.genes}
\title{A function to plot the heatmap of DE genes detected at a given FDR
threshold from the Meta-analysis.}
\usage{
heatmap.sig.genes(result, meta.method, fdr.cut, color = "GR")
}
\arguments{
\item{result}{is the output from MetaDE.}

\item{meta.method}{is the meta-analysis method used in MetaDE.}

\item{fdr.cut}{is the FDR cutoff used to select the DE genes.}

\item{color}{is the color of the heatmap.}
}
\value{
a figure shows the standardized expression levels for the DE genes
detected by meta analysis across studies/datasets.
}
\description{
Heatmap of selected DE genes
Author: Tianzhou Ma
Institution: University of pittsburgh
Date: 08/20/2016
The \code{heatmap.sig.genes} is a function to draw the Heatmap of DE genes
given a FDR cut point obtained from the Meta-analysis.
}
\examples{
\dontrun{
meta.method <- 'AW'
fdr.cut <- 1e-7
pdf('heatmap.test.pdf')
heatmap.sig.genes(meta.res.p, meta.method=meta.method,
                    fdr.cut=fdr.cut,color="GR")
dev.off()
}
}

