% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDE.pvalue.R
\name{MetaDE.pvalue}
\alias{MetaDE.pvalue}
\title{Meta analysis by combining p-value}
\usage{
MetaDE.pvalue(x, meta.method, rth = NULL, parametric = TRUE)
}
\arguments{
\item{x}{is a list with components:
\itemize{
\item{p:}{ a list of p values for each dataset.}
\item{bp:}{  a list of p values calculated from permutation for each dataset. 
This part can be NULL if you just have the p-values from your own method.}
}}

\item{meta.method}{is a character to specify the Meta-analysis method 
used to combine the p-values.}

\item{rth}{is the option for roP and roP.OC method. rth means the 
rth smallest p-value.}

\item{parametric}{is a logical values indicating whether the parametric 
methods is chosen to calculate the p-values in meta-analysis.}

\item{x}{is a list with components:}
}
\value{
a list with components: \cr
\itemize{
\item{stat:}{ a matrix with rows representing genes. It is the statistic for 
the selected meta analysis method of combining p-values.}
\item{pval:}{ the p-value from meta analysis for each gene for the above 
stat.}
\item{FDR:}{ the FDR of the p-value for each gene for the above stat.}
\item{AW.weight:}{ The optimal weight assigned to each dataset/study for 
each gene if the '\code{AW}' method was chosen.}
}
}
\description{
Meta analysis by combining p-value 
The \code{MetaDE} is a function to identify genes associated with the 
response/phenoype of interest (can be either group, continuous or survival) 
by combining p-values from multiple studies(datasets). 
The main input consists of p-values from your own method/calculations.
}
\examples{
data('Leukemia')
data('LeukemiaLabel')
data <- Leukemia
K <- length(data)
clin.data <- lapply(label, function(x) {data.frame(x)} )
for (k in 1:length(clin.data)){
 colnames(clin.data[[k]]) <- "label"
}
select.group <- c('inv(16)','t(15;17)')
ref.level <- "inv(16)"
data.type <- "continuous"
ind.method <- c('limma','limma','limma')
resp.type <- "twoclass"
paired <- rep(FALSE,length(data))
ind.res <- Indi.DE.Analysis(data=data,clin.data= clin.data, 
                        data.type=data.type,resp.type = resp.type,
                        response='label',
                        ind.method=ind.method,select.group = select.group,
                        ref.level=ref.level,paired=paired)
meta.method <- "AW"
meta.res <- MetaDE.pvalue(ind.res,meta.method,rth=NULL,parametric=TRUE)
summary <- data.frame(ind.p = meta.res$ind.p,
                      stat = meta.res$meta.analysis$stat,
                      pval = meta.res$meta.analysis$pval,
                      FDR = meta.res$meta.analysis$FDR,
                      weight = meta.res$meta.analysis$AW.weight)
}
