% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MetaAnalysisDE.R
\name{MetaDE}
\alias{MetaDE}
\title{Main Function for Meta DE analysis: microarray & RNAseq.}
\usage{
MetaDE(data, clin.data, data.type, resp.type, response, covariate, ind.method,
  meta.method, select.group, ref.level, paired = NULL, rth, AW.type, REM.type,
  asymptotic = FALSE, nperm = 100, tail = "abs", asymptotic.p = FALSE,
  seed = 12345, ...)
}
\arguments{
\item{data}{is a list of K elements, where K is the number of studies, each
element is a microarray or RNAseq expression matrix with G rows and N
columns, where G is number of matched genes and N is the sample size.}

\item{clin.data}{is a list of K elements, each element includes is a
clinical data frame with N rows and p columns, where N is the sample size
and p is the number of clinical variables (main response included).}

\item{data.type}{is a character indicating the data type of the elements
in \code{data}, must be "continuous" or "discrete".}

\item{resp.type}{is a character indicating the response type of the
\code{response} variable selected, must be one of "twoclass", "multiclass",
"continuous" and "survival".}

\item{response}{is one column name of \code{clin.data}, indicating the
phenotype of interest. For survival, two column names have to be specified,
the first is the survival time and the second is the censoring status.}

\item{covariate}{are the clinical covariates you wish to adjust for in the
DE analysis, can be a vector of column names or NULL.}

\item{ind.method}{is a character vector to specify the method used to test
if there is association between the gene expression and outcome variable.}

\item{meta.method}{is a character to specify the Meta-analysis method
used to combine the p-values, effect sizes or ranks.}

\item{paired}{is a logical vecter of size K to indicate whether the study
is paired design?}

\item{rth}{is the option for roP and roP.OC method. rth means the
rth smallest p-value.}

\item{AW.type}{is the option for "AW" method only, choose from
"original", "uncond" or "cond".}

\item{REM.type}{is the option for "REM" method only, choose from
"HS","HO", "DL", "SJ", "EB" or "RML".}

\item{asymptotic}{is a logical value indicating whether asymptotic
distribution should be used. If FALSE, permutation will be performed.}

\item{nperm}{is the number of permutations. Applicable when \code{asymptotic}
is FALSE.}

\item{tail}{is a character string specifying the alternative hypothesis,
must be one of "abs" (default), "low" or "high".}

\item{asymptotic.p}{is a logical values indicating whether the parametric
methods is chosen to calculate the p-values in meta-analysis.}

\item{select.group:}{for two-class comparison only, specify the two groups
for comparison when the group factor has more than two levels.}

\item{ref.level:}{for two-class/multi-class comparison only, specify the
reference level of the group factor.}

\item{seed:}{Optional initial seed for random number generator.}
}
\value{
a list with components: \cr
\itemize{
\item{stat}{a matrix with rows reprenting genes. It is the statistic for
the selected meta analysis method of combining p-values.}
\item{pval}{the p-value from meta analysis for each gene for the above
stat.}
\item{FDR}{the FDR of the p-value for each gene for the above stat.}
\item{AW.weight}{The optimal weight assigned to each dataset/study for
each gene if the '\code{AW}' method was chosen.}
}
}
\description{
Main Function for Meta analysis: microarray & RNAseq
Author: Tianzhou Ma
Institution: University of pittsburgh
Date: 08/20/2016
The \code{MetaDE} is a function to identify genes associated with the
response/phenoype of interest (can be either group, continuous or survival)
by integrating multiple studies(datasets).
}
\note{
multiclass available for limma, samr, edgeR, DESeq2 only;
}
\examples{
data('Leukemia')
data('LeukemiaLabel')
data <- Leukemia
K <- length(data)
clin.data <- lapply(label, function(x) {data.frame(x)} )
for (k in 1:length(clin.data)){
 colnames(clin.data[[k]]) <- "label"
}
select.group <- c('inv(16)','t(15;17)')
ref.level <- "inv(16)"
data.type <- "continuous"
ind.method <- c('limma','limma','sam')
resp.type <- "twoclass"
paired <- rep(FALSE,length(data))
meta.method <- "AW"
meta.res <- MetaDE(data=data,clin.data = clin.data,
                    data.type=data.type,resp.type = resp.type,
                    response='label',covariate = NULL,
                    ind.method=ind.method, meta.method=meta.method,
                    select.group = select.group, ref.level=ref.level,
                    paired=paired, rth=NULL,AW.type="original",
                    REM.type=NULL,tail='abs')
meta.method <- "Fisher"
meta.res <- MetaDE(data=data,clin.data = clin.data,
                    data.type=data.type,resp.type = resp.type,
                    response='label',covariate = NULL,
                    ind.method=ind.method, meta.method=meta.method,
                    select.group = select.group, ref.level=ref.level,
                    paired=paired, rth=NULL,AW.type="original",
                    REM.type=NULL,tail='abs')
meta.method <- "Fisher.OC"
meta.res <- MetaDE(data=data,clin.data = clin.data,
                    data.type=data.type,resp.type = resp.type,
                    response='label',covariate = NULL,
                    ind.method=ind.method, meta.method=meta.method,
                    select.group = select.group, ref.level=ref.level,
                    paired=paired, rth=NULL,AW.type="original",
                    REM.type=NULL,tail='high')
meta.method <- "FEM"
meta.res <- MetaDE(data=data,clin.data = clin.data,
                    data.type=data.type,resp.type = resp.type,
                    response='label',covariate = NULL,
                    ind.method=ind.method, meta.method=meta.method,
                    select.group = select.group, ref.level=ref.level,
                    paired=paired, rth=NULL,AW.type="original",
                    REM.type=NULL,tail='abs')
meta.method <- "REM"
REM.type <- "HO"
meta.res <- MetaDE(data=data,clin.data = clin.data,
                    data.type=data.type,resp.type = resp.type,
                    response='label',covariate = NULL,
                    ind.method=ind.method, meta.method=meta.method,
                    select.group = select.group, ref.level=ref.level,
                    paired=paired, rth=NULL,AW.type="original",
                    REM.type=REM.type,tail='abs')
meta.method <- "SR"
meta.res <- MetaDE(data=data,clin.data = clin.data,
                    data.type=data.type,resp.type = resp.type,
                    response='label',covariate = NULL,
                    ind.method=ind.method, meta.method=meta.method,
                    select.group = select.group, ref.level=ref.level,
                    paired=paired, rth=NULL,AW.type="original",
                    REM.type=NULL,tail='abs')
meta.method <- 'minMCC'
meta.res <- MetaDE(data=data,clin.data = clin.data,
                    data.type=data.type,resp.type = resp.type,
                    response='label',covariate = NULL,
                    ind.method=ind.method, meta.method=meta.method,
                    select.group = select.group, ref.level=ref.level,
                    paired=paired, rth=NULL,AW.type="original",
                    REM.type=NULL,tail='abs')
}

