% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{PathAnalysis}
\alias{PathAnalysis}
\title{Main Function for pathway analysis.}
\usage{
PathAnalysis(meta.p = NULL, pathway = c(Biocarta.genesets, GOBP.genesets,
  GOCC.genesets, GOMF.genesets, KEGG.genesets, Reactome.genesets),
  enrichment = c("KS", "Fisher's exact"), p.cut = NULL,
  DEgene.number = 200, size.min = 15, size.max = 500)
}
\arguments{
\item{meta.p}{is a vector of meta-analysis p-value.}

\item{pathway}{is a vector of pathway databases used for functional 
analysis, see data(pathways) for more details.}

\item{enrichment}{is the method used for pathway analysis, must be one of
"KS" and "Fisher's exact".}

\item{p.cut}{is the p-value cutoff to select the DE genes, option for 
Fisher's exact method only.}

\item{DEgene.number}{is the top number of DE genes, option for 
Fisher's exact method only.}

\item{size.min}{is the minimum pathway size to be included in the 
functional analysis.}

\item{size.max}{is the maximum pathway size to be included in the 
functional analysis.}
}
\value{
a data frame with columns: \cr
\itemize{
\item{pvalue}{ the p-value from pathway analysis for each pathway}
\item{qvalue}{ the q-value from pathway analysis for each pathway}
\item{OddsRatio}{ optional, the odds ratio from Fisher's exact test method}
\item{logOR}{ optional, the log odds ratio from Fisher's exact test method}
\item{DEgenes}{ optional, the set of DE genes in each pathway}
}
}
\description{
Main Function for pathway analysis
The \code{PathAnalysis} is a function to perform pathway analysis 
(a.k.a. gene set enrichment test) for functional annotation of a candidate
gene list or an ordered gene result from Meta DE analysis output.
}
\examples{
\dontrun{
meta.p  <- meta.res$meta.analysis$pval
ks.result <- PathAnalysis(meta.p = meta.p, enrichment = "KS")
fisher.result <- PathAnalysis(meta.p = meta.p, enrichment = "Fisher's exact")
}
}
