% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostHoc.R
\name{posthoc.aw}
\alias{posthoc.aw}
\title{Post-hoc analysis on AW results.}
\usage{
posthoc.aw(result)
}
\arguments{
\item{result}{is the output from MetaDE AW method}
}
\value{
a new AW result with additional result (the overall effect size
directionality) from the post-hoc analysis.
}
\description{
Post-hoc analysis on AW results
The \code{posthoc.aw} is a function to perform post-hoc analysis on AW 
results to determine the overall effect size directionality.
}
\examples{
\dontrun{
posthoc.result <- posthoc.aw(result=meta.res)
}
}
