% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryMeta.R
\name{summary.meta}
\alias{summary.meta}
\title{Function to summarize the meta-analysis results.}
\usage{
summary.meta(result, meta.method)
}
\arguments{
\item{result}{is the output from MetaDE}

\item{meta.method}{is the meta-analysis method used in MetaDE}
}
\value{
a summary table including individual study statistics and pvalue,
meta-analysis test statistics, pvalue, FDR, AW weights etc.
}
\description{
Function to summarize the results in tabular form 
The \code{summary.meta} is a function to summarize the meta-analysis 
results from the MetaDE output.
}
\examples{
\dontrun{
meta.method <- 'AW'
summary.result <- summary.meta(result=meta.res, 
                              meta.method = meta.method)
summary.posthoc.result <- summary.meta(result=posthoc.result, 
                              meta.method = meta.method)
}
}
